+++ workers/ohsome_quality_analyst/reports/building_report/report.py
@@ -1,0 +1,13 @@
+from ohsome_quality_analyst.base.report import BaseReport, IndicatorLayer
+
+
+class BuildingReport(BaseReport):
+    def set_indicator_layer(self):
+        self.indicator_layer = (
+            IndicatorLayer("MappingSaturation", "building_count"),
+            IndicatorLayer("Currentness", "building_count"),
+            IndicatorLayer("TagsRatio", "building_count"),
+        )
+
+    def combine_indicators(self) -> None:
+        super().combine_indicators()