+++ workers/tests/unittests/test_report_building_report.py
@@ -1,17 +1,17 @@
 import unittest
 from unittest.mock import Mock
 
-from ohsome_quality_analyst.reports.jrc_requirements.report import JrcRequirements
+from ohsome_quality_analyst.reports.building_report.report import BuildingReport
 
 from .utils import get_geojson_fixture
 
 
-class TestReportJrcRequirements(unittest.TestCase):
+class TestReportBuildingReport(unittest.TestCase):
     # TODO: Test case for indicator.result undefined
     def test_combine_indicators_mean(self):
 
         geometry = get_geojson_fixture("heidelberg-altstadt-geometry.geojson")
-        report = JrcRequirements(geometry)
+        report = BuildingReport(geometry)
         report.set_indicator_layer()
 
         # Mock indicator objects with a fixed result value